+++ app/common/services/grid/grid.js
@@ -1,7 +1,6 @@
 import angular from 'angular';
 
 import 'angular-ui-grid';
-import 'angular-ui-grid/ui-grid.css!';
 import './grid.css!';
 
 const name = 'projectX.grid';